+++ packages/impromat-api/src/dtos/inputs/element-input.ts
@@ -5,10 +5,16 @@ import {
   IntersectionType,
   PartialType,
 } from '@nestjs/graphql';
-import { MaxLength } from 'class-validator';
+import { IsIn, Length, MaxLength } from 'class-validator';
 import { Nullable } from 'src/utils/nullish';
 import { ElementVisibility } from '../types/element-visibility.dto';
 import { IdInput } from './id-input';
+
+@InputType()
+export class ElementTagConnectInput {
+  @Field(() => [IdInput])
+  connect: IdInput[];
+}
 
 @InputType()
 export class CreateElementInput {
@@ -23,8 +29,19 @@ export class CreateElementInput {
   @MaxLength(10000)
   markdown: Nullable<string>;
 
+  @Field(() => String, {
+    defaultValue: 'en',
+    description: 'Language code (e.g. en, de) of the element.',
+  })
+  @Length(2)
+  @IsIn(['en', 'de'])
+  languageCode: string;
+
   @Field(() => Int, { nullable: true })
   orderIndex?: number;
+
+  @Field(() => ElementTagConnectInput, { nullable: true })
+  tags: ElementTagConnectInput;
 }
 
 @InputType()