+++ packages/impromat/test/component/fixtures/community-dev-page.ts
@@ -1,0 +1,7 @@
+import { DevPage } from "./dev-page.js";
+
+export class CommunityDevPage extends DevPage {
+  async goto() {
+    await this.page.goto(`./community`);
+  }
+}