+++ packages/impromat-api/CHANGELOG.md
@@ -1,5 +1,18 @@
 # impromat-api
 
+## 0.9.0
+
+### Minor Changes
+
+- 67c31e4: Define language for custom elements
+- f2509b3: Add custom element tags for User created elements
+
+## 0.8.0
+
+### Minor Changes
+
+- 56ddd9c: Increase element search threshold to 0.9 to allow fuzzier search results
+
 ## 0.7.0
 
 ### Minor Changes